Fixed Documentation
